Fix unnecessary i64 cast in system determinism test helper
`AsteroidField::radius` is already i64; the `as i64` tripped
clippy::unnecessary_cast under --all-targets (test targets weren't in
the per-crate clippy runs the sprints used). Workspace clippy
--all-targets now clean.

+++ reachlock-core/src/generator/system.rs
@@ -546,7 +546,7 @@ mod tests {
         for f in &sys.asteroid_fields {
             checksum = checksum
                 .wrapping_mul(31)
-                .wrapping_add(f.center.x.0 ^ f.center.y.0 ^ f.radius as i64);
+                .wrapping_add(f.center.x.0 ^ f.center.y.0 ^ f.radius);
         }
         for s in &sys.station_slots {
             checksum = checksum
